Move MPI and PBLAS related replacements to `common/distr.py`
