+++ dace/frontend/common/__init__.py
@@ -1,2 +1,3 @@
 # Copyright 2019-2021 ETH Zurich and the DaCe authors. All rights reserved.
 from .einsum import create_einsum_sdfg
+from .distr import *
